requirements.txt
asyncio
requests
beautifulsoup4
discord.py
